pyproject.toml
[project]
name = "vehicle_asset_lib"
version = "0.1.0"
description = "Unattended Vehicle Asset Pickup Library"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "sqlalchemy>=2.0.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]

[project.scripts]
vehicle-asset = "vehicle_asset_lib.cli:app"

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["vehicle_asset_lib"]
